grails-login-app comlete

+++ grails-login-app/grails-app/controllers/grails/login/app/DepartmentController.groovy
@@ -24,16 +24,14 @@ class DepartmentController {
         if (department.validate()) {
             returnedResult = departmentService.createDepartment(department)
             def status, msg
-//            log.info('create(): {}', returnedResult)
             if (returnedResult) {
                 status = true
-                msg = 'department created'
+                msg = g.message(code: 'dept.created')
             } else if (!returnedResult) {
                 status = false
-                msg = 'department could not be created'
+                msg = g.message(code: 'dept.creation.failed')
 
             }
-//            render(view: 'createDepartment', model: [status: status, msg: msg])
             flash.message = [info: msg, success: status]
             redirect(controller: 'department', action: 'all')
             return
@@ -47,7 +45,6 @@ class DepartmentController {
             redirect(controller: 'auth', action: '403')
             return
         }
-//        log.info('edit(): {}', id)
         [department: departmentService.getDepartment(id)]
     }
 
@@ -59,15 +56,13 @@ class DepartmentController {
         def returnedResult
         returnedResult = departmentService.updateDepartment(params)
         def status, msg
-//        log.info('update(): {}', returnedResult)
         if (returnedResult) {
             status = true
-            msg = 'department updated'
+            msg = g.message(code: 'dept.updated')
         } else if (!returnedResult) {
             status = false
-            msg = 'department could not be updated'
+            msg = g.message(code: 'dept.not.updated')
         }
-//        render(view: 'all', model: [status: status, msg: msg, departments: departmentService.getAll()])
         flash.message = [info: msg, success: status]
         redirect(controller: 'department', action: 'all')
         return
@@ -78,18 +73,23 @@ class DepartmentController {
             redirect(controller: 'auth', action: '403')
             return
         }
-        def returnedResult
+        def returnedResult, status, msg
+        if(departmentService.getDepartment(id).students.size()>0){
+            status = false
+            msg = g.message(code: 'dept.has.zero')
+            flash.message = [info: msg, success: status]
+            redirect(controller: 'department', action: 'all')
+            return
+        }
         returnedResult = departmentService.deleteDepartment(id)
-        def status, msg
-//        log.info('delete(): {}', returnedResult)
+
         if (returnedResult) {
             status = true
-            msg = 'department deleted'
+            msg = g.message(code: 'dept.deleted')
         } else if (!returnedResult) {
             status = false
-            msg = 'department could not be deleted'
+            msg = g.message(code: 'dept.not.deleted')
         }
-//        render(view: 'all', model: [status: status, msg: msg, departments: departmentService.getAll()])
         flash.message = [info: msg, success: status]
         redirect(controller: 'department', action: 'all')
         return
